fix #847 When downloading more than 2GB, assertionError may occur

+++ javamelody-core/src/main/java/net/bull/javamelody/CounterRequestMXBean.java
@@ -132,7 +132,7 @@ public interface CounterRequestMXBean {
 			return request.getSystemErrorPercentage();
 		}
 
-		public int getResponseSizeMean() {
+		public long getResponseSizeMean() {
 			return request.getResponseSizeMean();
 		}
 	}
