+++ javamelody-core/src/main/java/net/bull/javamelody/internal/common/JavaLogger.java
@@ -64,7 +64,7 @@ class JavaLogger implements JavaMelodyLogger {
 	/** {@inheritDoc} */
 	@Override
 	public void logHttpRequest(HttpServletRequest httpRequest, String requestName, long duration,
-			boolean systemError, int responseSize, String loggerName) {
+			boolean systemError, long responseSize, String loggerName) {
 		final java.util.logging.Logger logger = java.util.logging.Logger.getLogger(loggerName);
 		if (logger.isLoggable(Level.INFO)) {
 			logger.info(LOG.buildLogMessage(httpRequest, duration, systemError, responseSize));
