+++ javamelody-core/src/main/java/net/bull/javamelody/internal/common/LogbackLogger.java
@@ -63,7 +63,7 @@ class LogbackLogger implements JavaMelodyLogger {
 	/** {@inheritDoc} */
 	@Override
 	public void logHttpRequest(HttpServletRequest httpRequest, String requestName, long duration,
-			boolean systemError, int responseSize, String loggerName) {
+			boolean systemError, long responseSize, String loggerName) {
 		final Logger logger = org.slf4j.LoggerFactory.getLogger(loggerName);
 		if (logger.isInfoEnabled()) {
 			logger.info(LOG.buildLogMessage(httpRequest, duration, systemError, responseSize));
