+++ javamelody-core/src/main/java/net/bull/javamelody/internal/model/CounterRequestContext.java
@@ -274,7 +274,7 @@ public class CounterRequestContext implements ICounterRequestContext, Cloneable,
 
 	@SuppressWarnings("unused")
 	void addChildRequest(Counter childCounter, String request, String requestId, long duration,
-			boolean systemError, int responseSize) {
+			boolean systemError, long responseSize) {
 		// si je suis le counter fils du counter du contexte parent
 		// comme sql pour http alors on ajoute la requête fille
 		if (parentContext != null && parentCounter.getName()
