+++ javamelody-core/src/main/java/net/bull/javamelody/internal/web/CounterResponseStream.java
@@ -27,7 +27,7 @@ import javax.servlet.http.HttpServletResponse;
  * @author Emeric Vernat
  */
 public class CounterResponseStream extends FilterServletOutputStream {
-	private int dataLength;
+	private long dataLength;
 
 	/**
 	 * Construit un servlet output stream associé avec la réponse spécifiée.
@@ -48,9 +48,9 @@ public class CounterResponseStream extends FilterServletOutputStream {
 
 	/**
 	 * Retourne la valeur de la propriété dataLength.
-	 * @return int
+	 * @return long
 	 */
-	public int getDataLength() {
+	public long getDataLength() {
 		return dataLength;
 	}
